Move enums to new package and use id value instead of ordinal
