+++ app/src/main/java/io/github/deprec8/enigmadroid/data/LoadingRepository.kt
@@ -23,7 +23,7 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.intPreferencesKey
-import io.github.deprec8.enigmadroid.data.objects.LoadingState
+import io.github.deprec8.enigmadroid.data.enums.LoadingState
 import io.github.deprec8.enigmadroid.data.objects.PreferencesKeys
 import io.github.deprec8.enigmadroid.data.source.local.devices.DevicesDatabase
 import io.github.deprec8.enigmadroid.data.source.network.NetworkDataSource
@@ -49,7 +49,7 @@ class LoadingRepository @Inject constructor(
     init {
         CoroutineScope(Dispatchers.Default).launch {
             dataStore.edit { preferences ->
-                preferences[loadingStateKey] = LoadingState.LOADING.ordinal
+                preferences[loadingStateKey] = LoadingState.LOADING.id
             }
         }
     }
@@ -68,19 +68,19 @@ class LoadingRepository @Inject constructor(
         if (currentLoadingState == LoadingState.LOADING || forceUpdate) {
             if (currentLoadingState != LoadingState.LOADING) {
                 dataStore.edit { preferences ->
-                    preferences[loadingStateKey] = LoadingState.LOADING.ordinal
+                    preferences[loadingStateKey] = LoadingState.LOADING.id
                 }
             }
 
             if (devicesDatabase.deviceDao().getAll().firstOrNull().isNullOrEmpty().not()) {
                 if (networkDataSource.isDeviceOnline()) {
                     dataStore.edit { preferences ->
-                        preferences[loadingStateKey] = LoadingState.LOADED.ordinal
+                        preferences[loadingStateKey] = LoadingState.LOADED.id
                     }
                 }
             } else {
                 dataStore.edit { preferences ->
-                    preferences[loadingStateKey] = LoadingState.NO_DEVICE_AVAILABLE.ordinal
+                    preferences[loadingStateKey] = LoadingState.NO_DEVICE_AVAILABLE.id
                 }
             }
         }
