+++ app/src/main/java/io/github/deprec8/enigmadroid/data/enums/ApiType.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.data.objects
+package io.github.deprec8.enigmadroid.data.enums
 
 enum class ApiType {
     RADIO, TV
