+++ app/src/main/java/io/github/deprec8/enigmadroid/data/enums/LoadingState.kt
@@ -17,8 +17,8 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.data.objects
+package io.github.deprec8.enigmadroid.data.enums
 
-enum class LoadingState {
-    LOADING, NO_DEVICE_AVAILABLE, DEVICE_NOT_ONLINE, LOADED,
+enum class LoadingState(val id: Int) {
+    LOADING(0), NO_DEVICE_AVAILABLE(1), DEVICE_NOT_ONLINE(2), LOADED(3),
 }
