+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/timers/TimersViewModel.kt
@@ -29,7 +29,7 @@ import io.github.deprec8.enigmadroid.data.ApiRepository
 import io.github.deprec8.enigmadroid.data.LoadingRepository
 import io.github.deprec8.enigmadroid.data.SearchHistoryRepository
 import io.github.deprec8.enigmadroid.data.SettingsRepository
-import io.github.deprec8.enigmadroid.data.objects.LoadingState
+import io.github.deprec8.enigmadroid.data.enums.LoadingState
 import io.github.deprec8.enigmadroid.model.ServiceList
 import io.github.deprec8.enigmadroid.model.Timer
 import io.github.deprec8.enigmadroid.model.TimerList
