+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/timers/timers.kt
@@ -73,7 +73,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.github.deprec8.enigmadroid.R
-import io.github.deprec8.enigmadroid.data.objects.LoadingState
+import io.github.deprec8.enigmadroid.data.enums.LoadingState
 import io.github.deprec8.enigmadroid.data.objects.TimerState
 import io.github.deprec8.enigmadroid.model.Timer
 import io.github.deprec8.enigmadroid.model.menu.MenuItem
