+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/tvEPG/TVEPGViewModel.kt
@@ -29,8 +29,8 @@ import io.github.deprec8.enigmadroid.data.ApiRepository
 import io.github.deprec8.enigmadroid.data.LoadingRepository
 import io.github.deprec8.enigmadroid.data.SearchHistoryRepository
 import io.github.deprec8.enigmadroid.data.SettingsRepository
-import io.github.deprec8.enigmadroid.data.objects.ApiType
-import io.github.deprec8.enigmadroid.data.objects.LoadingState
+import io.github.deprec8.enigmadroid.data.enums.ApiType
+import io.github.deprec8.enigmadroid.data.enums.LoadingState
 import io.github.deprec8.enigmadroid.model.EPGEvent
 import io.github.deprec8.enigmadroid.model.EPGEventList
 import io.github.deprec8.enigmadroid.utils.FilterUtils
